+++ formidable/classes/controllers/FrmEntriesController.php
@@ -121,7 +121,9 @@ class FrmEntriesController {
 					$col_id .= '-_-form' . $form_col->form_id;
 				}
 
-				if ( isset($form_col->field_options['separate_value']) && $form_col->field_options['separate_value'] ) {
+				$has_separate_value = ! FrmField::is_option_empty( $form_col, 'separate_value' );
+				$is_not_post_status = FrmField::is_option_empty( $form_col, 'post_field' );
+				if ( $has_separate_value && $is_not_post_status ) {
 					$columns[ $form_id . '_frmsep_' . $col_id ] = FrmAppHelper::truncate( $form_col->name, 35 );
 				}
 				$columns[ $form_id . '_' . $col_id ] = FrmAppHelper::truncate( $form_col->name, 35 );
@@ -512,10 +514,6 @@ class FrmEntriesController {
 	public static function filter_shortcode_value( $value, $tag, $atts, $field ) {
         $plain_text = add_filter('frm_plain_text_email', true);
 		FrmEntryFormat::textarea_display_value( $field->type, $plain_text, $value );
-
-        if ( isset($atts['show']) && $atts['show'] == 'value' ) {
-            return $value;
-        }
 
         return $value;
     }