+++ types/library/toolset/types/embedded/bootstrap.php
@@ -157,7 +157,7 @@ function wpcf_embedded_init() {
 	
 	// Toolset Forms
 	if ( ! defined( 'WPTOOLSET_FORMS_VERSION' ) ) {
-		$toolset_common_bootstrap = Toolset_Common_Bootstrap::getInstance();
+		$toolset_common_bootstrap = Toolset_Common_Bootstrap::get_instance();
 		$toolset_common_sections = array(
 			'toolset_forms'
 		);
@@ -190,14 +190,6 @@ function wpcf_embedded_init() {
     define( 'WPCF_EMBEDDED_RES_RELPATH', WPCF_EMBEDDED_RELPATH . '/resources' );
 
     // TODO INCLUDES!
-    //
-    // Please add all required includes here
-    // Since Types 1.2 we can consider existing code as core.
-    // All new functionalities should be added as includes HERE
-    // and marked with @since Types $version.
-    //
-    // Thanks!
-    //
 
     // Basic
     /*
@@ -243,25 +235,16 @@ function wpcf_embedded_init() {
         require_once WPCF_EMBEDDED_INC_ABSPATH . '/cred.php';
     }
 
-    /*
-     *
-     *
-     * TODO This is a must for now.
-     * See if any fields need to be loaded.
-     *
-     * 1. Checkboxes - may be missing when submitted
-     */
+    // This is a must for now.
+	// See if any fields need to be loaded.
+	// 1. Checkboxes - may be missing when submitted
     require_once WPCF_EMBEDDED_INC_ABSPATH . '/fields/checkbox.php';
 
+    // Use this call to load basic scripts and styles if necesary
+	// wpcf_enqueue_scripts();
+	require_once WPCF_EMBEDDED_ABSPATH . '/usermeta-init.php';
 
-    /*
-     *
-     *
-     * Use this call to load basic scripts and styles if necesary
-     * wpcf_enqueue_scripts();
-     */
-    require_once WPCF_EMBEDDED_ABSPATH . '/usermeta-init.php';
-    // Include frontend or admin code
+	// Include frontend or admin code
     if ( is_admin() ) {
 
         require_once WPCF_EMBEDDED_ABSPATH . '/admin.php';
@@ -279,7 +262,6 @@ function wpcf_embedded_init() {
 
     global $wpcf;
 
-    // TODO since Types 1.2 Continue adding new functionalities HERE
     /*
      * Consider code already there as core.
      * Use hooks to add new functionalities